requirements.txt
langchain_core==0.2.16
langchain_openai==0.1.15
libclang==18.1.1
numpy==1.24.2
openai==1.35.13
python_tlsh==4.5.0
regex==2022.10.31
Requests==2.32.3
scikit_learn==1.4.2
tlsh==0.2.0
